# userservices/.optic/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Optic Generated OpenAPI",
    "version": "1.0.0"
  },
  "paths": {
    "/api/aggregator/getDashboardForAggregator": {
      "get": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "repaidInvoiceCount": {
                          "type": "number"
                        },
                        "UtilizedAmount": {
                          "type": "number"
                        },
                        "Totalpendinginvoices": {
                          "type": "number"
                        },
                        "Totalinvoices": {
                          "type": "number"
                        },
                        "Totalhospital": {
                          "type": "number"
                        },
                        "TotalfundedInvoices": {
                          "type": "number"
                        },
                        "TotalRejectedinvoices": {
                          "type": "number"
                        },
                        "TotalRejectedClaimAmount": {
                          "type": "number"
                        },
                        "TotalPendingClaimAmount": {
                          "type": "number"
                        },
                        "TotalInvoicesClaimAmount": {
                          "type": "number"
                        },
                        "TotalFundedClaimamount": {
                          "type": "number"
                        },
                        "TotalAvailableLimit": {
                          "type": "number"
                        },
                        "Repayment": {
                          "type": "number"
                        },
                        "CreditLimit": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "success",
                        "repaidInvoiceCount",
                        "UtilizedAmount",
                        "Totalpendinginvoices",
                        "Totalinvoices",
                        "Totalhospital",
                        "TotalfundedInvoices",
                        "TotalRejectedinvoices",
                        "TotalRejectedClaimAmount",
                        "TotalPendingClaimAmount",
                        "TotalInvoicesClaimAmount",
                        "TotalFundedClaimamount",
                        "TotalAvailableLimit",
                        "Repayment",
                        "CreditLimit"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "errors"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/aggregator/getInvoiceGraphToAggregator": {
      "get": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "Totalinvoicescount": {
                          "type": "number"
                        },
                        "TotalcountRepaid": {
                          "type": "number"
                        },
                        "TotalcountRejected": {
                          "type": "number"
                        },
                        "TotalcountInProcess": {
                          "type": "number"
                        },
                        "TotalcountDisbursed": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "success",
                        "Totalinvoicescount",
                        "TotalcountRepaid",
                        "TotalcountRejected",
                        "TotalcountInProcess",
                        "TotalcountDisbursed"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/aggregator/getSecondPieGraphForAggregator": {
      "get": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "TotalInvoices": {
                          "type": "number"
                        },
                        "TotalFunded": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "success",
                        "TotalInvoices",
                        "TotalFunded"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/aggregator/splitForAgri": {
      "get": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "HospitalROI": {
                    "type": "string"
                  },
                  "HospitalLoan": {
                    "type": "string"
                  },
                  "AggregetorLoan": {
                    "type": "string"
                  },
                  "AggregeorId": {
                    "type": "string"
                  }
                },
                "required": [
                  "HospitalROI",
                  "HospitalLoan",
                  "AggregetorLoan",
                  "AggregeorId"
                ]
              }
            }
          }
        },
        "responses": {
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "errors"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "HospitalROI": {
                    "type": "string"
                  },
                  "HospitalLoan": {
                    "type": "string"
                  },
                  "AggregetorLoan": {
                    "type": "string"
                  },
                  "AggregeorId": {
                    "type": "string"
                  }
                },
                "required": [
                  "HospitalROI",
                  "HospitalLoan",
                  "AggregetorLoan",
                  "AggregeorId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "errors"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/aggregator/getAllInvoicebyHospitalToAggregator": {
      "get": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "message": {
                          "type": "array",
                          "items": {}
                        },
                        "count": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "success",
                        "message",
                        "count"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}